feat: support for external dependencies for the KRM KCL spec
